adicionando politicas para controlar sessões e possibilidades

+++ app/Http/Controllers/CategoriaController.php
@@ -14,11 +14,13 @@ class CategoriaController extends Controller
 
     public function index()
     {
+        $this->authorize('viewAny', Categoria::class);
         return view('categoria.index')->with('categorias', Categoria::all());
     }
 
     public function destroy($id)
     {
+        $this->authorize('delete', Categoria::find($id));
         $categoria = Categoria::find($id);
         if($categoria->delete()){
             return back()->with('mensagem', "Categoria excluída");
